backend_fastapi/src/app/authmain/util/authmain_util.py
from datetime import timedelta, datetime
from fastapi import HTTPException, status
from jose import jwt
from src.app.config.authkey import SECRET_KEY, ALGORITHM
from typing import Annotated
from fastapi import Depends
from src.app.config.authkey import oauth2_bearer


def create_access_token(email: str, user_id: int, expires_delta: timedelta) -> str:
    encode = {'sub': email, 'id': user_id}
    expires = datetime.utcnow() + expires_delta
    encode.update({'exp': expires})
    return jwt.encode(encode, SECRET_KEY, algorithm=ALGORITHM)


async def get_current_user(token: Annotated[str, Depends(oauth2_bearer)]) -> dict:
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        email: str = payload.get('sub')
        user_id: int = payload.get('id')
        if email is None or user_id is None:
            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
                                detail='Could not validate user')
        return {'email': email, 'id': user_id}
    except JWTError:
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
                            detail='Could not validate')
